convert query function added for showing SELECT statements in questions list function working well
