+++ server_python/server.py
@@ -4,6 +4,7 @@ import server_python.sv_general
 import server_python.sv_answers
 from server_python.config import app
 
+
 if __name__ == "__main__":
     app.run(
         debug=True,
